Move collaborator validation
